:recycle: files reorganization
