+++ ozkour-back/config/routes/talk.js
@@ -1,7 +1,7 @@
 
 const { getTalk } = require('../../domain/talks-sheet')
 
-const { createSlides } = require('../../google-api/slideService')
+const { createSlides } = require('../../domain/slideService')
 
 module.exports = [
   {
