+++ ozkour-back/domain/slideDataOrganizer.js
