+++ ozkour-back/domain/talks-sheet.js
@@ -2,7 +2,7 @@ const dayjs = require('dayjs')
 const customParseFormat = require('dayjs/plugin/customParseFormat')
 dayjs.extend(customParseFormat)
 const sheetsWrapper = require('../infrastructure/googlesheets/sheetWrapper')
-const { getIdOfTalkFileByYear } = require('../google-api/drive')
+const { getIdOfTalkFileByYear } = require('../infrastructure/googledrive/drive')
 const utils = require('../Utils/dateUtils')
 const isSameOrAfter = require('dayjs/plugin/isSameOrAfter')
 const isSameOrBefore = require('dayjs/plugin/isSameOrBefore')
